Updated deleteThing function.

+++ backend/controllers/stuff.js
@@ -52,7 +52,19 @@ exports.updateThing = (req, res, next) => {
 
 // Delete a thing
 exports.deleteThing = (req, res, next) => {
-  Thing.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: "Deleted!" }))
-    .catch((err) => res.status(400).json({ error: err }));
+  Thing.findOne({ _id: req.params.id }).then((thing) => {
+    if (!thing) {
+      return res.status(404).json({
+        error: new Error("No such thing!"),
+      });
+    }
+    if (thing.userId != req.auth.userId) {
+      return res.status(401).json({
+        error: new Error("Unauthorized request!"),
+      });
+    }
+    Thing.deleteOne({ _id: req.params.id })
+      .then(() => res.status(200).json({ message: "Deleted!" }))
+      .catch((err) => res.status(400).json({ error: err }));
+  });
 }
